fixing some bugs and adding notification related models

+++ commerces/models.py
@@ -26,7 +26,7 @@ class Commerce(models.Model):
     latitude = models.FloatField(blank=True, null=True)
     longitude = models.FloatField(blank=True, null=True)
     ratingCount = models.IntegerField(default=0)
-    ratingTotal = models.IntegerField(default=0)
+    ratingTotal = models.FloatField(default=0)
     profilePicture = models.URLField(max_length=300, blank=True, null=True)
     headerPicture = models.URLField(max_length=300, blank=True, null=True)
     softDelete = models.DateTimeField(blank=True, null=True)
